Vit/requirements.txt
jupyter 
transformers 
datasets
torchvision
Pillow
fastapi
uvicorn